First version of a minimal playable game

+++ src/models/hit_target.cr
@@ -11,9 +11,16 @@ class HitTarget < ApplicationRecord
 
   model_action :hit, game.grid do
     def model_action_controller
-      model.db.exec "DELETE FROM #{model.table_name} WHERE id=#{model.id}"
+      return unless model
+
+      if game_player = model.game.game_players.find { |gp| gp.session_id == ctx.session.id.to_s }
+        model.db.exec "DELETE FROM #{model.table_name} WHERE id=#{model.id}"
+        game_player.score = game_player.score.value + 1
+        game_player.save
+      end
 
       model_template.turbo_stream.to_html(ctx.response)
+      Crumble::Turbo::ModelTemplateRefreshService.notify(model.game.leaderboard)
     end
   end
 end
